Move scratch_notify_topic_subscribers to tasks.py.
Hopefully gets it registered.
