+++ djangobb_forum/subscription.py
@@ -6,8 +6,8 @@ from celery.decorators import task
 
 from djangobb_forum import settings as forum_settings
 from djangobb_forum.util import absolute_url
+from djangobb_forum.tasks import scratch_notify_topic_subscribers
 from djangobb_forum.models import Post
-
 
 
 if "mailer" in settings.INSTALLED_APPS:
@@ -64,21 +64,3 @@ def email_topic_subscribers(post):
             }
         #html_content = html_version(post)
         send_mail(subject, text_content, settings.DEFAULT_FROM_EMAIL, [to_email])
-
-
-@task 
-def scratch_notify_topic_subscribers(post_id):
-    """
-    Scratch task for notifying subscribers to a topic that someone has made a
-    new post.
-    """
-    from notifications.models import SocialAction
-    post = Post.objects.select_related('topic').get(id=post_id)
-    topic = post.topic
-    if post != topic.head:
-        social_action = SocialAction(
-            actor = post.user,
-            object = topic,
-        )
-        social_action.save()
-        # save() adds recipients and sends out notifications
